feat: UI组件重构和功能优化
