+++ .env.example
@@ -22,6 +22,9 @@ SEED_SUPER_ADMIN_NAME="诗扶小筑管理员"
 CORS_ORIGIN=""
 UPLOAD_SESSION_TTL_MINUTES="45"
 
+# Custom uploads storage path (absolute path, e.g., /var/www/huangshifu-wiki/uploads)
+UPLOADS_PATH=""
+
 # Database backup
 BACKUP_PASSWORD="replace_with_backup_password"
 BACKUP_RETAIN_COUNT="20"
