+++ config/server.config.env.example
@@ -35,6 +35,9 @@ COOKIE_SECURE="true"
 UPLOAD_SESSION_TTL_MINUTES="45"
 EDIT_LOCK_TTL_MINUTES="20"
 
+# Custom uploads storage path (absolute path, e.g., /var/www/huangshifu-wiki/uploads)
+UPLOADS_PATH=""
+
 # =============================
 # LLM / Gemini config (frontend build-time)
 # =============================
