+++ src/components/Modal/index.ts
@@ -1,0 +1,2 @@
+export { ConfirmModal } from "./ConfirmModal";
+export { FormModal } from "./FormModal";
