+++ src/lib/i18n.ts
@@ -9,22 +9,25 @@ const locales: Record<string, Locale> = {
   academy: academyLocale,
 };
 
-function getNestedValue(obj: any, keyPath: string): string {
-  return keyPath.split('.').reduce((acc, key) => acc?.[key], obj) || keyPath;
+function getNestedValue(obj: Record<string, unknown>, keyPath: string): string {
+  return keyPath.split('.').reduce((acc: unknown, key: string) => {
+    if (acc && typeof acc === 'object') return (acc as Record<string, unknown>)[key];
+    return undefined;
+  }, obj) as string || keyPath;
 }
 
-function replaceParams(text: string, params: Record<string, any>): string {
+function replaceParams(text: string, params: Record<string, string | number>): string {
   return text.replace(/\{\{([^}]+)\}\}/g, (match, key) => {
-    return params[key.trim()] || match;
+    return String(params[key.trim()] ?? match);
   });
 }
 
 export function useI18n() {
   const { theme } = useTheme();
   
-  const t = (key: string, params?: Record<string, any>): string => {
+  const t = (key: string, params?: Record<string, string | number>): string => {
     const locale = locales[theme] || defaultLocale;
-    const text = getNestedValue(locale, key);
+    const text = getNestedValue(locale as unknown as Record<string, unknown>, key);
     return params ? replaceParams(text, params) : text;
   };
   
@@ -34,8 +37,8 @@ export function useI18n() {
 export function getI18n(theme: string) {
   const locale = locales[theme] || defaultLocale;
   
-  const t = (key: string, params?: Record<string, any>): string => {
-    const text = getNestedValue(locale, key);
+  const t = (key: string, params?: Record<string, string | number>): string => {
+    const text = getNestedValue(locale as unknown as Record<string, unknown>, key);
     return params ? replaceParams(text, params) : text;
   };
