+++ src/pages/Search.tsx
@@ -25,6 +25,8 @@ import { apiGet, apiUpload } from "../lib/apiClient";
 import { toDateValue } from "../lib/dateUtils";
 import { useTheme } from "../context/ThemeContext";
 import { withThemeSearch, mergeSearchParamsWithTheme } from "../lib/theme";
+import type { ThemeName } from "../lib/theme";
+import type { WikiItem, PostItem, GalleryItem, SongItem, AlbumItem } from "../types/entities";
 
 type SearchSuggestion = {
 	type: "keyword" | "wiki" | "post" | "music" | "album";
@@ -33,18 +35,288 @@ type SearchSuggestion = {
 	id?: string;
 };
 
+interface SearchWikiCardProps {
+	page: WikiItem;
+	viewMode: string;
+	theme: ThemeName;
+}
+
+const SearchWikiCard = React.memo(({ page, viewMode, theme }: SearchWikiCardProps) => (
+	<Link
+		to={withThemeSearch(`/wiki/${page.slug}`, theme)}
+		className={clsx(
+			viewMode === "list"
+				? "flex gap-4 p-4 bg-white rounded-xl border border-gray-100 hover:border-brand-olive/20 hover:shadow-lg transition-all w-full"
+				: "bg-white p-6 rounded-3xl border border-gray-100 hover:border-brand-olive/20 hover:shadow-lg transition-all group",
+		)}
+	>
+		{viewMode === "list" ? (
+			<>
+				<div className="w-24 h-24 bg-brand-cream/50 rounded-lg flex items-center justify-center flex-shrink-0">
+					<Book
+						size={32}
+						className="text-brand-olive/40"
+					/>
+				</div>
+				<div className="flex-1 min-w-0">
+					<div className="flex items-center gap-2 mb-1">
+						<span className="px-2 py-0.5 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
+							{page.category}
+						</span>
+					</div>
+					<h3 className="text-lg font-serif font-bold mb-1 group-hover:text-brand-olive transition-colors truncate">
+						{page.title}
+					</h3>
+					<p className="text-gray-400 text-sm line-clamp-2 italic">
+						{page.content
+							.replace(/[#*`]/g, "")
+							.substring(0, 100)}
+					</p>
+					<p className="text-gray-300 text-xs mt-1">
+						{page.content
+							.replace(/[#*`]/g, "")
+							.substring(0, 50)}
+						...
+					</p>
+				</div>
+			</>
+		) : (
+			<>
+				<div className="flex items-center gap-2 mb-3">
+					<span className="px-2 py-0.5 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
+						{page.category}
+					</span>
+				</div>
+				<h3 className="text-xl font-serif font-bold mb-2 group-hover:text-brand-olive transition-colors">
+					{page.title}
+				</h3>
+				<p className="text-gray-400 text-sm line-clamp-2 mb-4 italic leading-relaxed">
+					{page.content
+						.replace(/[#*`]/g, "")
+						.substring(0, 100)}
+					...
+				</p>
+				<div className="flex items-center justify-between text-gray-400 text-[10px]">
+					<span className="flex items-center gap-1">
+						<Clock size={12} />{" "}
+						{toDateValue(page.updatedAt)
+							? format(
+									toDateValue(page.updatedAt)!,
+									"yyyy-MM-dd",
+								)
+							: "刚刚"}
+					</span>
+					<ChevronRight
+						size={14}
+						className="group-hover:translate-x-1 transition-transform"
+					/>
+				</div>
+			</>
+		)}
+	</Link>
+));
+
+interface SearchGalleryCardProps {
+	gallery: GalleryItem;
+	viewMode: string;
+	theme: ThemeName;
+}
+
+const SearchGalleryCard = React.memo(({ gallery, viewMode, theme }: SearchGalleryCardProps) => (
+	<Link
+		to={withThemeSearch(`/gallery/${gallery.id}`, theme)}
+		className={clsx(
+			viewMode === "list"
+				? "flex gap-4 p-3 bg-white rounded-xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all w-full"
+				: "bg-white rounded-2xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all group",
+		)}
+	>
+		{viewMode === "list" ? (
+			<>
+				<div className="w-24 h-24 bg-gray-100 rounded-lg overflow-hidden flex-shrink-0">
+					<SmartImage
+						src={gallery.images[0]?.url}
+						alt=""
+						className="w-full h-full object-cover"
+					/>
+				</div>
+				<div className="flex-1 min-w-0 flex items-center">
+					<div className="flex-1 min-w-0">
+						<h3 className="text-sm font-serif font-bold truncate group-hover:text-brand-olive transition-colors">
+							{gallery.title}
+						</h3>
+						<p className="text-xs text-gray-400">
+							{gallery.description || "暂无描述"}
+						</p>
+						<p className="text-[10px] text-gray-300 mt-1">
+							{(gallery.description || "").substring(
+								0,
+								50,
+							)}
+							...
+						</p>
+					</div>
+					<span className="text-xs text-gray-400 bg-gray-100 px-2 py-1 rounded-full flex-shrink-0">
+						{gallery.images.length} 张
+					</span>
+				</div>
+			</>
+		) : (
+			<>
+				<div className="h-32 overflow-hidden">
+					<SmartImage
+						src={gallery.images[0]?.url}
+						alt=""
+						className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
+					/>
+				</div>
+				<div className="p-4">
+					<h3 className="text-sm font-serif font-bold truncate group-hover:text-brand-olive transition-colors">
+						{gallery.title}
+					</h3>
+					<p className="text-[10px] text-gray-400">
+						{gallery.images.length} 张图片
+					</p>
+				</div>
+			</>
+		)}
+	</Link>
+));
+
+interface SearchMusicCardProps {
+	track: SongItem;
+	viewMode: string;
+	theme: ThemeName;
+}
+
+const SearchMusicCard = React.memo(({ track, viewMode, theme }: SearchMusicCardProps) => (
+	<Link
+		to={withThemeSearch(`/music/${track.id}`, theme)}
+		className={clsx(
+			viewMode === "list"
+				? "flex gap-4 p-4 bg-white rounded-xl border border-gray-100 hover:border-pink-200 hover:shadow-lg transition-all w-full"
+				: "bg-white p-6 rounded-3xl border border-gray-100 hover:border-pink-200 hover:shadow-lg transition-all group",
+		)}
+	>
+		{viewMode === "list" ? (
+			<>
+				<div className="w-16 h-16 rounded-lg overflow-hidden bg-brand-cream flex-shrink-0">
+					<SmartImage
+						src={track.cover}
+						alt=""
+						className="w-full h-full object-cover"
+					/>
+				</div>
+				<div className="flex-1 min-w-0 flex items-center">
+					<div className="flex-1 min-w-0">
+						<h3 className="text-sm font-serif font-bold truncate group-hover:text-pink-500 transition-colors">
+							{track.title}
+						</h3>
+						<p className="text-xs text-gray-500 truncate">
+							{track.artist} — {track.album}
+						</p>
+					</div>
+				</div>
+			</>
+		) : (
+			<div className="flex items-center gap-4">
+				<div className="w-16 h-16 rounded-xl overflow-hidden bg-brand-cream flex-shrink-0">
+					<SmartImage
+						src={track.cover}
+						alt=""
+						className="w-full h-full object-cover"
+					/>
+				</div>
+				<div className="flex-1 min-w-0">
+					<h3 className="text-lg font-serif font-bold truncate group-hover:text-pink-500 transition-colors">
+						{track.title}
+					</h3>
+					<p className="text-sm text-gray-500 truncate">
+						{track.artist}
+					</p>
+					<p className="text-xs text-gray-400 truncate">
+						{track.album}
+					</p>
+				</div>
+			</div>
+		)}
+	</Link>
+));
+
+interface SearchAlbumCardProps {
+	album: AlbumItem;
+	viewMode: string;
+	theme: ThemeName;
+}
+
+const SearchAlbumCard = React.memo(({ album, viewMode, theme }: SearchAlbumCardProps) => (
+	<Link
+		to={withThemeSearch(`/album/${album.id}`, theme)}
+		className={clsx(
+			viewMode === "list"
+				? "flex gap-4 p-3 bg-white rounded-xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all w-full"
+				: "bg-white rounded-2xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all group",
+		)}
+	>
+		{viewMode === "list" ? (
+			<>
+				<div className="w-16 h-16 rounded-lg overflow-hidden bg-gray-100 flex-shrink-0">
+					<SmartImage
+						src={album.cover}
+						alt=""
+						className="w-full h-full object-cover"
+					/>
+				</div>
+				<div className="flex-1 min-w-0 flex items-center">
+					<div className="flex-1 min-w-0">
+						<h3 className="text-sm font-serif font-bold truncate group-hover:text-purple-500 transition-colors">
+							{album.title}
+						</h3>
+						<p className="text-xs text-gray-400">
+							{album.artist}
+						</p>
+					</div>
+					<span className="text-xs text-gray-400 bg-gray-100 px-2 py-1 rounded-full flex-shrink-0">
+						{album.trackCount} 曲
+					</span>
+				</div>
+			</>
+		) : (
+			<>
+				<div className="h-40 overflow-hidden">
+					<SmartImage
+						src={album.cover}
+						alt=""
+						className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
+					/>
+				</div>
+				<div className="p-4">
+					<h3 className="text-sm font-serif font-bold truncate group-hover:text-purple-500 transition-colors">
+						{album.title}
+					</h3>
+					<p className="text-[10px] text-gray-400">
+						{album.artist} · {album.trackCount} 曲
+					</p>
+				</div>
+			</>
+		)}
+	</Link>
+));
+
+interface SearchResults {
+		wiki: WikiItem[];
+		posts: PostItem[];
+		galleries: GalleryItem[];
+		music: SongItem[];
+		albums: AlbumItem[];
+	}
+
 const Search = () => {
 	const [searchParams, setSearchParams] = useSearchParams();
 	const navigate = useNavigate();
 	const initialQuery = searchParams.get("q") || "";
 	const [searchQuery, setSearchQuery] = useState(initialQuery);
-	const [results, setResults] = useState<{
-		wiki: any[];
-		posts: any[];
-		galleries: any[];
-		music: any[];
-		albums: any[];
-	}>({ wiki: [], posts: [], galleries: [], music: [], albums: [] });
+	const [results, setResults] = useState<SearchResults>({ wiki: [], posts: [], galleries: [], music: [], albums: [] });
 	const [loading, setLoading] = useState(false);
 	const [hasSearched, setHasSearched] = useState(Boolean(initialQuery));
 	const [activeTab, setActiveTab] = useState<
@@ -637,79 +909,12 @@ const Search = () => {
 											)}
 										>
 											{results.wiki.map((page) => (
-												<Link
+												<SearchWikiCard
 													key={page.id}
-													to={withThemeSearch(`/wiki/${page.slug}`, theme)}
-													className={clsx(
-														viewMode === "list"
-															? "flex gap-4 p-4 bg-white rounded-xl border border-gray-100 hover:border-brand-olive/20 hover:shadow-lg transition-all w-full"
-															: "bg-white p-6 rounded-3xl border border-gray-100 hover:border-brand-olive/20 hover:shadow-lg transition-all group",
-													)}
-												>
-													{viewMode === "list" ? (
-														<>
-															<div className="w-24 h-24 bg-brand-cream/50 rounded-lg flex items-center justify-center flex-shrink-0">
-																<Book
-																	size={32}
-																	className="text-brand-olive/40"
-																/>
-															</div>
-															<div className="flex-1 min-w-0">
-																<div className="flex items-center gap-2 mb-1">
-																	<span className="px-2 py-0.5 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
-																		{page.category}
-																	</span>
-																</div>
-																<h3 className="text-lg font-serif font-bold mb-1 group-hover:text-brand-olive transition-colors truncate">
-																	{page.title}
-																</h3>
-																<p className="text-gray-400 text-sm line-clamp-2 italic">
-																	{page.content
-																		.replace(/[#*`]/g, "")
-																		.substring(0, 100)}
-																</p>
-																<p className="text-gray-300 text-xs mt-1">
-																	{page.content
-																		.replace(/[#*`]/g, "")
-																		.substring(0, 50)}
-																	...
-																</p>
-															</div>
-														</>
-													) : (
-														<>
-															<div className="flex items-center gap-2 mb-3">
-																<span className="px-2 py-0.5 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
-																	{page.category}
-																</span>
-															</div>
-															<h3 className="text-xl font-serif font-bold mb-2 group-hover:text-brand-olive transition-colors">
-																{page.title}
-															</h3>
-															<p className="text-gray-400 text-sm line-clamp-2 mb-4 italic leading-relaxed">
-																{page.content
-																	.replace(/[#*`]/g, "")
-																	.substring(0, 100)}
-																...
-															</p>
-															<div className="flex items-center justify-between text-gray-400 text-[10px]">
-																<span className="flex items-center gap-1">
-																	<Clock size={12} />{" "}
-																	{toDateValue(page.updatedAt)
-																		? format(
-																				toDateValue(page.updatedAt)!,
-																				"yyyy-MM-dd",
-																			)
-																		: "刚刚"}
-																</span>
-																<ChevronRight
-																	size={14}
-																	className="group-hover:translate-x-1 transition-transform"
-																/>
-															</div>
-														</>
-													)}
-												</Link>
+													page={page}
+													viewMode={viewMode}
+													theme={theme}
+												/>
 											))}
 										</div>
 									</motion.section>
@@ -775,65 +980,12 @@ const Search = () => {
 											)}
 										>
 											{results.galleries.map((gallery) => (
-												<Link
+												<SearchGalleryCard
 													key={gallery.id}
-													to={withThemeSearch(`/gallery/${gallery.id}`, theme)}
-													className={clsx(
-														viewMode === "list"
-															? "flex gap-4 p-3 bg-white rounded-xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all w-full"
-															: "bg-white rounded-2xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all group",
-													)}
-												>
-													{viewMode === "list" ? (
-														<>
-															<div className="w-24 h-24 bg-gray-100 rounded-lg overflow-hidden flex-shrink-0">
-																<SmartImage
-																	src={gallery.images[0]?.url}
-																	alt=""
-																	className="w-full h-full object-cover"
-																/>
-															</div>
-															<div className="flex-1 min-w-0 flex items-center">
-																<div className="flex-1 min-w-0">
-																	<h3 className="text-sm font-serif font-bold truncate group-hover:text-brand-olive transition-colors">
-																		{gallery.title}
-																	</h3>
-																	<p className="text-xs text-gray-400">
-																		{gallery.description || "暂无描述"}
-																	</p>
-																	<p className="text-[10px] text-gray-300 mt-1">
-																		{(gallery.description || "").substring(
-																			0,
-																			50,
-																		)}
-																		...
-																	</p>
-																</div>
-																<span className="text-xs text-gray-400 bg-gray-100 px-2 py-1 rounded-full flex-shrink-0">
-																	{gallery.images.length} 张
-																</span>
-															</div>
-														</>
-													) : (
-														<>
-															<div className="h-32 overflow-hidden">
-																<SmartImage
-																	src={gallery.images[0]?.url}
-																	alt=""
-																	className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
-																/>
-															</div>
-															<div className="p-4">
-																<h3 className="text-sm font-serif font-bold truncate group-hover:text-brand-olive transition-colors">
-																	{gallery.title}
-																</h3>
-																<p className="text-[10px] text-gray-400">
-																	{gallery.images.length} 张图片
-																</p>
-															</div>
-														</>
-													)}
-												</Link>
+													gallery={gallery}
+													viewMode={viewMode}
+													theme={theme}
+												/>
 											))}
 										</div>
 									</motion.section>
@@ -858,58 +1010,12 @@ const Search = () => {
 											)}
 										>
 											{results.music.map((track) => (
-												<Link
+												<SearchMusicCard
 													key={track.docId}
-													to={withThemeSearch(`/music/${track.id}`, theme)}
-													className={clsx(
-														viewMode === "list"
-															? "flex gap-4 p-4 bg-white rounded-xl border border-gray-100 hover:border-pink-200 hover:shadow-lg transition-all w-full"
-															: "bg-white p-6 rounded-3xl border border-gray-100 hover:border-pink-200 hover:shadow-lg transition-all group",
-													)}
-												>
-													{viewMode === "list" ? (
-														<>
-															<div className="w-16 h-16 rounded-lg overflow-hidden bg-brand-cream flex-shrink-0">
-																<SmartImage
-																	src={track.cover}
-																	alt=""
-																	className="w-full h-full object-cover"
-																/>
-															</div>
-															<div className="flex-1 min-w-0 flex items-center">
-																<div className="flex-1 min-w-0">
-																	<h3 className="text-sm font-serif font-bold truncate group-hover:text-pink-500 transition-colors">
-																		{track.title}
-																	</h3>
-																	<p className="text-xs text-gray-500 truncate">
-																		{track.artist} — {track.album}
-																	</p>
-																</div>
-															</div>
-														</>
-													) : (
-														<div className="flex items-center gap-4">
-															<div className="w-16 h-16 rounded-xl overflow-hidden bg-brand-cream flex-shrink-0">
-																<SmartImage
-																	src={track.cover}
-																	alt=""
-																	className="w-full h-full object-cover"
-																/>
-															</div>
-															<div className="flex-1 min-w-0">
-																<h3 className="text-lg font-serif font-bold truncate group-hover:text-pink-500 transition-colors">
-																	{track.title}
-																</h3>
-																<p className="text-sm text-gray-500 truncate">
-																	{track.artist}
-																</p>
-																<p className="text-xs text-gray-400 truncate">
-																	{track.album}
-																</p>
-															</div>
-														</div>
-													)}
-												</Link>
+													track={track}
+													viewMode={viewMode}
+													theme={theme}
+												/>
 											))}
 										</div>
 									</motion.section>
@@ -934,58 +1040,12 @@ const Search = () => {
 											)}
 										>
 											{results.albums.map((album) => (
-												<Link
+												<SearchAlbumCard
 													key={album.docId}
-													to={withThemeSearch(`/album/${album.id}`, theme)}
-													className={clsx(
-														viewMode === "list"
-															? "flex gap-4 p-3 bg-white rounded-xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all w-full"
-															: "bg-white rounded-2xl border border-gray-100 overflow-hidden hover:shadow-lg transition-all group",
-													)}
-												>
-													{viewMode === "list" ? (
-														<>
-															<div className="w-16 h-16 rounded-lg overflow-hidden bg-gray-100 flex-shrink-0">
-																<SmartImage
-																	src={album.cover}
-																	alt=""
-																	className="w-full h-full object-cover"
-																/>
-															</div>
-															<div className="flex-1 min-w-0 flex items-center">
-																<div className="flex-1 min-w-0">
-																	<h3 className="text-sm font-serif font-bold truncate group-hover:text-purple-500 transition-colors">
-																		{album.title}
-																	</h3>
-																	<p className="text-xs text-gray-400">
-																		{album.artist}
-																	</p>
-																</div>
-																<span className="text-xs text-gray-400 bg-gray-100 px-2 py-1 rounded-full flex-shrink-0">
-																	{album.tracksCount} 曲
-																</span>
-															</div>
-														</>
-													) : (
-														<>
-															<div className="h-40 overflow-hidden">
-																<SmartImage
-																	src={album.cover}
-																	alt=""
-																	className="w-full h-full object-cover group-hover:scale-110 transition-transform duration-500"
-																/>
-															</div>
-															<div className="p-4">
-																<h3 className="text-sm font-serif font-bold truncate group-hover:text-purple-500 transition-colors">
-																	{album.title}
-																</h3>
-																<p className="text-[10px] text-gray-400">
-																	{album.artist} · {album.tracksCount} 曲
-																</p>
-															</div>
-														</>
-													)}
-												</Link>
+													album={album}
+													viewMode={viewMode}
+													theme={theme}
+												/>
 											))}
 										</div>
 									</motion.section>
