+++ src/pages/Wiki.tsx
@@ -1,8 +1,6 @@
 import React, {
 	useEffect,
 	useState,
-	useRef,
-	useCallback,
 	useMemo,
 } from "react";
 import {
@@ -24,7 +22,6 @@ import {
 	Edit3,
 	Plus,
 	ChevronRight,
-	Search,
 	Tag,
 	Clock,
 	User as UserIcon,
@@ -38,11 +35,9 @@ import {
 	GitBranch,
 	Network,
 	MapPin,
-	Heart,
 	ThumbsDown,
 	ThumbsUp,
 	Pin,
-	Image as ImageIcon,
 } from "lucide-react";
 import { useUserPreferences } from "../context/UserPreferencesContext";
 import { useTheme } from "../context/ThemeContext";
@@ -50,50 +45,32 @@ import { ViewModeSelector } from "../components/ViewModeSelector";
 import { VIEW_MODE_CONFIG } from "../lib/viewModes";
 import { clsx } from "clsx";
 import { motion, AnimatePresence } from "motion/react";
-import { summarizeWikiContent, generateWikiIntro } from "../services/aiService";
-import { uploadMarkdownImage } from "../services/imageService";
+import { summarizeWikiContent } from "../services/aiService";
 import { useToast } from "../components/Toast";
 import { copyToClipboard, toAbsoluteInternalUrl } from "../lib/copyLink";
-import { apiDelete, apiGet, apiPost, apiPut } from "../lib/apiClient";
+import { apiDelete, apiGet, apiPost } from "../lib/apiClient";
 import {
 	ContentStatus,
 	getStatusText,
 	splitTagsInput,
 } from "../lib/contentUtils";
-import { formatDate, toDateValue } from "../lib/dateUtils";
-import { randomId } from "../lib/randomId";
+import { formatDate } from "../lib/dateUtils";
 import { withThemeSearch } from "../lib/theme";
-import MdEditor from "react-markdown-editor-lite";
-import MarkdownIt from "markdown-it";
-import "react-markdown-editor-lite/lib/index.css";
 import WikiLinkPreview from "../components/WikiLinkPreview";
 import RelationGraph, {
 	RelationGraphData,
 	WikiRelationType as GraphRelationType,
 } from "../components/wiki/RelationGraph";
-import { LocationTagInput } from "../components/LocationTagInput";
 import Pagination from "../components/Pagination";
+import WikiCard from "../components/wiki/WikiCard";
+import WikiEditorComponent from "../components/wiki/WikiEditor";
+import type {
+	WikiItemWithRelations,
+	WikiRelationRecord,
+} from "../components/wiki/types";
+import { RELATION_TYPE_LABELS } from "../components/wiki/types";
 
 const DEFAULT_PAGE_SIZE = 24;
-
-const mdParser = new MarkdownIt({
-	html: true,
-	linkify: true,
-	typographer: true,
-});
-
-type WikiRelationType =
-	| "related_person"
-	| "work_relation"
-	| "timeline_relation"
-	| "custom";
-
-type WikiRelationRecord = {
-	type: WikiRelationType;
-	targetSlug: string;
-	label?: string;
-	bidirectional: boolean;
-};
 
 type WikiRelationResolved = WikiRelationRecord & {
 	typeLabel: string;
@@ -102,13 +79,6 @@ type WikiRelationResolved = WikiRelationRecord & {
 	inferred: boolean;
 	sourceSlug: string;
 	sourceTitle: string;
-};
-
-const RELATION_TYPE_LABELS: Record<WikiRelationType, string> = {
-	related_person: "相关人物",
-	work_relation: "作品关联",
-	timeline_relation: "时间线关联",
-	custom: "自定义关系",
 };
 
 const academyWikiLecturers = [
@@ -137,33 +107,7 @@ const academyWikiCopyMappings = [
 	},
 ];
 
-type WikiItem = {
-	id: string;
-	slug: string;
-	title: string;
-	category: string;
-	content: string;
-	tags?: string[];
-	eventDate?: string | null;
-	status?: ContentStatus;
-	reviewNote?: string | null;
-	reviewedBy?: string | null;
-	reviewedAt?: string | null;
-	favoritesCount?: number;
-	favoritedByMe?: boolean;
-	likesCount?: number;
-	dislikesCount?: number;
-	likedByMe?: boolean;
-	dislikedByMe?: boolean;
-	isPinned?: boolean;
-	lastEditorUid: string;
-	lastEditorName: string;
-	relations?: WikiRelationRecord[];
-	locationCode?: string | null;
-	locationName?: string | null;
-	createdAt: string;
-	updatedAt: string;
-};
+type WikiItem = WikiItemWithRelations;
 
 type WikiBranchStatus =
 	| "draft"
@@ -577,138 +521,12 @@ const WikiList = () => {
 						)}
 					>
 						{paginatedPages.map((page) => (
-							<div
+							<WikiCard
 								key={page.id}
-								className={clsx(
-									"relative group",
-									viewMode === "list" && "flex",
-								)}
-							>
-								<Link
-									to={withThemeSearch(`/wiki/${page.slug}`, theme)}
-									className={clsx(
-										viewMode === "list"
-											? "flex gap-4 p-4 bg-white rounded-xl border hover:border-brand-olive/20 hover:shadow-lg transition-all w-full"
-											: "block bg-white p-8 rounded-[32px] border hover:border-brand-olive/20 hover:shadow-xl transition-all",
-										page.isPinned && viewMode !== "list"
-											? "border-l-4 border-l-brand-olive"
-											: "border-gray-100",
-										page.isPinned &&
-											viewMode === "list" &&
-											"border-l-4 border-l-brand-olive",
-									)}
-								>
-									{viewMode === "list" ? (
-										<>
-											<div className="w-24 h-24 bg-brand-cream/50 rounded-lg flex items-center justify-center flex-shrink-0">
-												<Book size={32} className="text-brand-olive/40" />
-											</div>
-											<div className="flex-1 min-w-0">
-												<div className="flex items-center gap-2 mb-1">
-													{page.isPinned && (
-														<span className="flex items-center gap-1 px-2 py-0.5 bg-brand-primary/10 text-brand-primary text-[10px] font-bold uppercase tracking-wider rounded">
-															<Pin size={8} /> 已置顶
-														</span>
-													)}
-													<span className="px-2 py-0.5 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
-														{page.category === "biography"
-															? "人物介绍"
-															: page.category === "music"
-																? "音乐作品"
-																: page.category === "album"
-																	? "专辑一览"
-																	: page.category === "timeline"
-																		? "时间轴"
-																		: page.category === "event"
-																			? "活动记录"
-																			: page.category}
-													</span>
-												</div>
-												<h3 className="text-lg font-serif font-bold mb-1 group-hover:text-brand-olive transition-colors truncate">
-													{page.title}
-												</h3>
-												<p className="text-gray-400 text-sm line-clamp-2 italic">
-													{page.content.replace(/[#*`]/g, "").substring(0, 100)}
-												</p>
-												<p className="text-gray-300 text-xs mt-1">
-													{page.content.replace(/[#*`]/g, "").substring(0, 50)}
-													...
-												</p>
-												<div className="flex items-center gap-3 text-gray-400 text-xs mt-2">
-													<span className="flex items-center gap-1">
-														<Clock size={10} />{" "}
-														{formatDate(page.updatedAt, "yyyy-MM-dd")}
-													</span>
-													<span className="flex items-center gap-1">
-														<Heart size={10} /> {page.likesCount || 0}
-													</span>
-												</div>
-											</div>
-										</>
-									) : (
-										<>
-											<div className="flex items-center gap-2 mb-4">
-												{page.isPinned && (
-													<span className="flex items-center gap-1 px-2 py-1 bg-brand-primary/10 text-brand-primary text-[10px] font-bold uppercase tracking-wider rounded">
-														<Pin size={10} /> 已置顶
-													</span>
-												)}
-												<span className="px-2 py-1 bg-brand-cream text-brand-olive text-[10px] font-bold uppercase tracking-wider rounded">
-													{page.category === "biography"
-														? "人物介绍"
-														: page.category === "music"
-															? "音乐作品"
-															: page.category === "album"
-																? "专辑一览"
-																: page.category === "timeline"
-																	? "时间轴"
-																	: page.category === "event"
-																		? "活动记录"
-																		: page.category}
-												</span>
-											</div>
-											<h3 className="text-2xl font-serif font-bold mb-4 group-hover:text-brand-olive transition-colors">
-												{page.title}
-											</h3>
-											<p className="text-gray-400 text-sm line-clamp-2 mb-6 italic leading-relaxed">
-												{page.content.replace(/[#*`]/g, "").substring(0, 100)}
-												...
-											</p>
-											<div className="flex items-center justify-between text-gray-400 text-xs">
-												<div className="flex items-center gap-3">
-													<span className="flex items-center gap-1">
-														<Clock size={12} />{" "}
-														{formatDate(page.updatedAt, "yyyy-MM-dd")}
-													</span>
-													<span className="flex items-center gap-1">
-														<Heart size={12} /> {page.likesCount || 0}
-													</span>
-													<span className="flex items-center gap-1">
-														<ThumbsDown size={12} /> {page.dislikesCount || 0}
-													</span>
-												</div>
-												<ChevronRight
-													size={16}
-													className="group-hover:translate-x-1 transition-transform"
-												/>
-											</div>
-										</>
-									)}
-								</Link>
-								<button
-									onClick={(event) => handleCopyWikiLink(event, page.slug)}
-									className={clsx(
-										"p-2 rounded-full border bg-white/90 text-gray-400 shadow-sm hover:text-brand-olive hover:border-brand-olive/30 transition-all",
-										viewMode === "list"
-											? "absolute top-4 right-4"
-											: "absolute bottom-5 right-5 sm:opacity-0 sm:group-hover:opacity-100",
-									)}
-									title="复制内链"
-									aria-label="复制百科内链"
-								>
-									<Link2 size={14} />
-								</button>
-							</div>
+								page={page}
+								viewMode={viewMode}
+								onCopyLink={handleCopyWikiLink}
+							/>
 						))}
 					</div>
 					{totalWikiPages > 1 && (
@@ -2168,657 +1986,7 @@ const WikiPullRequestDetail = () => {
 	);
 };
 
-const WikiEditor = () => {
-	const { slug } = useParams();
-	const isNew = !slug || slug === "new";
-	const navigate = useNavigate();
-	const { user, profile, isAdmin, isBanned } = useAuth();
-	const { theme } = useTheme();
-
-	const [formData, setFormData] = useState({
-		title: "",
-		slug: "",
-		category: "biography",
-		content: "",
-		tags: "",
-		eventDate: "",
-		relations: [] as WikiRelationRecord[],
-		locationCode: "",
-		locationName: "",
-	});
-	const [savingMode, setSavingMode] = useState<"draft" | "pending" | null>(
-		null,
-	);
-	const [generating, setGenerating] = useState(false);
-	const { show } = useToast();
-
-	const [newRelation, setNewRelation] = useState<WikiRelationRecord>({
-		type: "related_person",
-		targetSlug: "",
-		label: "",
-		bidirectional: false,
-	});
-
-	type RelationSearchSuggestion = {
-		type: "keyword" | "wiki" | "post" | "music" | "album";
-		text: string;
-		subtext?: string;
-		id?: string;
-	};
-
-	const [relationSearchResults, setRelationSearchResults] = useState<
-		RelationSearchSuggestion[]
-	>([]);
-	const [relationSearchLoading, setRelationSearchLoading] = useState(false);
-	const [showRelationDropdown, setShowRelationDropdown] = useState(false);
-	const [relationSelectedIndex, setRelationSelectedIndex] = useState(-1);
-	const relationSearchRef = useRef<HTMLDivElement>(null);
-	const relationSearchTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-
-	const parseInternalLink = (input: string): { slug: string; displayText: string } | null => {
-		const match = input.match(/^\[\[([^\]|]+)(?:\|([^\]]+))?\]\]$/);
-		if (match) {
-			return {
-				slug: (match[2] || match[1]).trim(),
-				displayText: match[1].trim(),
-			};
-		}
-		return null;
-	};
-
-	const handleRelationInputChange = useCallback((input: string) => {
-		const internalLink = parseInternalLink(input);
-		if (internalLink) {
-			const shouldSetLabel = !newRelation.label && internalLink.displayText !== internalLink.slug;
-			setNewRelation((prev) => ({
-				...prev,
-				targetSlug: internalLink.slug,
-				label: shouldSetLabel ? internalLink.displayText : prev.label,
-			}));
-			setRelationSearchResults([]);
-			setShowRelationDropdown(false);
-			return;
-		}
-
-		setNewRelation((prev) => ({
-			...prev,
-			targetSlug: input,
-		}));
-
-		if (relationSearchTimeoutRef.current) {
-			clearTimeout(relationSearchTimeoutRef.current);
-		}
-		if (!input || input.length < 2) {
-			setRelationSearchResults([]);
-			setShowRelationDropdown(false);
-			return;
-		}
-		relationSearchTimeoutRef.current = setTimeout(async () => {
-			setRelationSearchLoading(true);
-			try {
-				const data = await apiGet<{ suggestions: RelationSearchSuggestion[] }>(
-					"/api/search/suggest",
-					{ q: input },
-				);
-				const wikiResults =
-					data.suggestions?.filter((s) => s.type === "wiki") || [];
-				setRelationSearchResults(wikiResults);
-				setShowRelationDropdown(wikiResults.length > 0);
-				setRelationSelectedIndex(-1);
-			} catch (e) {
-				console.error("Relation search error:", e);
-			} finally {
-				setRelationSearchLoading(false);
-			}
-		}, 300);
-	}, [newRelation.label]);
-
-	const searchWikiRelations = useCallback((q: string) => {
-		handleRelationInputChange(q);
-	}, [handleRelationInputChange]);
-
-	useEffect(() => {
-		const handleClickOutside = (e: MouseEvent) => {
-			if (
-				relationSearchRef.current &&
-				!relationSearchRef.current.contains(e.target as Node)
-			) {
-				setShowRelationDropdown(false);
-			}
-		};
-		document.addEventListener("mousedown", handleClickOutside);
-		return () => document.removeEventListener("mousedown", handleClickOutside);
-	}, []);
-
-	useEffect(() => {
-		if (!isNew) {
-			const fetchPage = async () => {
-				try {
-					const response = await apiGet<{ page: WikiItem }>(
-						`/api/wiki/${slug}`,
-					);
-					const data = response.page;
-					setFormData({
-						title: data.title,
-						slug: data.slug,
-						category: data.category,
-						content: data.content,
-						tags: data.tags?.join(", ") || "",
-						eventDate: data.eventDate || "",
-						relations: (data.relations as WikiRelationRecord[]) || [],
-						locationCode: data.locationCode || "",
-						locationName: data.locationName || "",
-					});
-				} catch (error) {
-					console.error("Error fetching wiki page for edit:", error);
-				}
-			};
-			fetchPage();
-		}
-	}, [slug, isNew]);
-
-	const handleLocationChange = (locationName: string, locationCode: string) => {
-		setFormData({ ...formData, locationName, locationCode });
-	};
-
-	const handleLocationClear = () => {
-		setFormData({ ...formData, locationName: "", locationCode: "" });
-	};
-
-	const handleSubmit = async (status: "draft" | "pending") => {
-		if (!user) return;
-		if (isBanned) {
-			show("账号已被封禁，无法编辑百科", { variant: "error" });
-			return;
-		}
-
-		if (formData.category === "music" && !isAdmin) {
-			show("只有管理员可以修改音乐分类的内容", { variant: "error" });
-			return;
-		}
-
-		if (!formData.title.trim()) {
-			show("请填写标题（*为必填项）", { variant: "error" });
-			return;
-		}
-		if (!formData.category) {
-			show("请选择分类（*为必填项）", { variant: "error" });
-			return;
-		}
-		if (!formData.content.trim()) {
-			show("请填写内容（*为必填项）", { variant: "error" });
-			return;
-		}
-
-		const pageSlug = (
-			isNew ? formData.slug || formData.title : slug || formData.slug
-		)
-			?.trim()
-			.toLowerCase()
-			.replace(/[\\/]/g, "-")
-			.replace(/\s+/g, "-");
-
-		if (!pageSlug) {
-			show("请先填写标题以生成页面标识", { variant: "error" });
-			return;
-		}
-
-		setSavingMode(status);
-
-		const pageData: any = {
-			title: formData.title,
-			slug: pageSlug,
-			category: formData.category,
-			content: formData.content,
-			tags: formData.tags
-				.split(",")
-				.map((t) => t.trim())
-				.filter((t) => t),
-			eventDate: formData.eventDate,
-			relations: formData.relations,
-			locationCode: formData.locationCode || null,
-			locationName: formData.locationName || null,
-			status,
-			lastEditorUid: user.uid,
-			lastEditorName: profile?.displayName || user.displayName || "匿名用户",
-		};
-
-		try {
-			if (isNew) {
-				try {
-					await apiGet<{ page: WikiItem }>(`/api/wiki/${pageSlug}`);
-					show("该页面标识已存在，请修改标题后重试", { variant: "error" });
-					setSavingMode(null);
-					return;
-				} catch {
-					// continue when page does not exist
-				}
-
-				const existingByTitle = await apiGet<{ pages: WikiItem[] }>(
-					"/api/wiki",
-					{
-						category: "all",
-					},
-				);
-				const duplicatedTitle = (existingByTitle.pages || []).some(
-					(item) => item.title.trim() === formData.title.trim(),
-				);
-
-				if (duplicatedTitle) {
-					show("该标题的百科已存在，请修改标题或编辑已有页面", {
-						variant: "error",
-					});
-					setSavingMode(null);
-					return;
-				}
-
-				await apiPost("/api/wiki", pageData);
-			} else {
-				await apiPut(`/api/wiki/${pageSlug}`, pageData);
-			}
-
-			await apiPost(`/api/wiki/${pageSlug}/revisions`, {
-				id: randomId(),
-				pageSlug,
-				title: formData.title,
-				content: formData.content,
-				editorUid: user.uid,
-				editorName: profile?.displayName || user.displayName || "匿名用户",
-			});
-
-			navigate(withThemeSearch(`/wiki/${pageSlug}`, theme));
-		} catch (e) {
-			console.error("Error saving wiki page:", e);
-			show("保存失败，请检查网络或权限", { variant: "error" });
-		}
-		setSavingMode(null);
-	};
-
-	return (
-		<div className="max-w-5xl mx-auto px-4 py-12">
-			<div className="bg-white rounded-[40px] p-8 sm:p-16 border border-gray-100 shadow-sm">
-				<div className="flex justify-between items-center mb-12">
-					<h1 className="text-4xl font-serif font-bold text-brand-olive">
-						{isNew ? "创建新百科" : "编辑百科"}
-					</h1>
-					<button
-						onClick={() => navigate(-1)}
-						className="p-2 text-gray-400 hover:text-red-500"
-					>
-						<X size={24} />
-					</button>
-				</div>
-
-				<form
-					onSubmit={(e) => {
-						e.preventDefault();
-						handleSubmit("pending");
-					}}
-					className="space-y-8"
-				>
-					<div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-						<div className="space-y-2">
-							<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-								标题
-							</label>
-							<input
-								type="text"
-								required
-								value={formData.title}
-								onChange={(e) =>
-									setFormData({ ...formData, title: e.target.value })
-								}
-								placeholder="例如：黄诗扶"
-								className="w-full px-6 py-4 bg-brand-cream rounded-2xl border-none focus:ring-2 focus:ring-brand-olive/20 font-serif text-xl"
-							/>
-						</div>
-						<div className="space-y-2">
-							<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-								分类
-							</label>
-							<select
-								value={formData.category}
-								onChange={(e) =>
-									setFormData({ ...formData, category: e.target.value })
-								}
-								className="w-full px-6 py-4 bg-brand-cream rounded-2xl border-none focus:ring-2 focus:ring-brand-olive/20 font-serif text-xl appearance-none"
-							>
-								<option value="biography">人物介绍</option>
-								<option value="music">音乐作品</option>
-								<option value="album">专辑一览</option>
-								<option value="timeline">时间线</option>
-								<option value="event">活动记录</option>
-							</select>
-						</div>
-						<div className="space-y-2">
-							<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-								事件日期 (可选)
-							</label>
-							<input
-								type="date"
-								value={formData.eventDate}
-								onChange={(e) =>
-									setFormData({ ...formData, eventDate: e.target.value })
-								}
-								className="w-full px-6 py-4 bg-brand-cream rounded-2xl border-none focus:ring-2 focus:ring-brand-olive/20 font-serif text-xl"
-							/>
-						</div>
-					</div>
-
-					<div className="space-y-2">
-						<div className="flex justify-between items-center">
-							<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-								内容 (Markdown) <span className="text-red-500">*</span>
-							</label>
-							<button
-								type="button"
-								onClick={async () => {
-									if (!formData.title)
-										return show("请先输入标题", { variant: "error" });
-									setGenerating(true);
-									const intro = await generateWikiIntro(formData.title);
-									if (intro)
-										setFormData({
-											...formData,
-											content: intro + "\n\n" + formData.content,
-										});
-									setGenerating(false);
-								}}
-								disabled={generating}
-								className="text-xs font-bold text-brand-olive flex items-center gap-1 hover:underline disabled:opacity-50"
-							>
-								<Sparkles size={12} />{" "}
-								{generating ? "生成中..." : "AI 辅助写开头"}
-							</button>
-						</div>
-						<div className="border border-gray-100 rounded-[32px] overflow-hidden">
-							<MdEditor
-								style={{ height: "500px" }}
-								renderHTML={(text) => {
-									const processed = text.replace(
-										/\[\[([^\]|]+)(?:\|([^\]]+))?\]\]/g,
-										(match, p1, p2) => {
-											const display = p1.trim();
-											const slug = p2 ? p2.trim() : p1.trim();
-											return `[${display}](/wiki/${slug})`;
-										},
-									);
-									return mdParser.render(processed);
-								}}
-								value={formData.content}
-								onChange={({ text }) =>
-									setFormData({ ...formData, content: text })
-								}
-								onImageUpload={uploadMarkdownImage}
-								placeholder="在这里输入百科内容，支持 Markdown 语法..."
-								config={{
-									view: {
-										menu: true,
-										md: true,
-										html: false,
-									},
-									canView: {
-										menu: true,
-										md: true,
-										html: true,
-										fullScreen: true,
-										hideMenu: false,
-									},
-								}}
-							/>
-						</div>
-					</div>
-
-					<div className="space-y-2">
-						<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-							标签 (逗号分隔)
-						</label>
-						<input
-							type="text"
-							value={formData.tags}
-							onChange={(e) =>
-								setFormData({ ...formData, tags: e.target.value })
-							}
-							placeholder="例如：古风, 原创, 歌手"
-							className="w-full px-6 py-4 bg-brand-cream rounded-2xl border-none focus:ring-2 focus:ring-brand-olive/20"
-						/>
-					</div>
-
-					<div className="space-y-2">
-						<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-							地点
-						</label>
-						<LocationTagInput
-							value={formData.locationName || null}
-							locationCode={formData.locationCode || null}
-							onChange={handleLocationChange}
-							onClear={handleLocationClear}
-						/>
-					</div>
-
-					<div className="space-y-4">
-						<div className="flex items-center justify-between">
-							<label className="text-xs font-bold uppercase tracking-widest text-brand-olive/60">
-								相关页面
-							</label>
-							<button
-								type="button"
-								onClick={() => {
-									if (!newRelation.targetSlug.trim()) {
-										show("请输入目标页面标识", { variant: "error" });
-										return;
-									}
-									setFormData({
-										...formData,
-										relations: [
-											...formData.relations,
-											{
-												...newRelation,
-												targetSlug: newRelation.targetSlug.trim(),
-											},
-										],
-									});
-									setNewRelation({
-										type: "related_person",
-										targetSlug: "",
-										label: "",
-										bidirectional: false,
-									});
-								}}
-								className="px-4 py-2 bg-brand-primary/10 text-brand-primary rounded-xl text-xs font-bold hover:bg-brand-primary/20 transition-all"
-							>
-								+ 添加关联
-							</button>
-						</div>
-
-						{formData.relations.length > 0 && (
-							<div className="space-y-2">
-								{formData.relations.map((relation, index) => (
-									<div
-										key={index}
-										className="flex items-center gap-3 p-3 bg-brand-cream/50 rounded-xl"
-									>
-										<span className="text-xs text-brand-primary font-bold min-w-[80px]">
-											{RELATION_TYPE_LABELS[relation.type]}
-										</span>
-										<span className="flex-1 text-sm truncate">
-											{relation.label || relation.targetSlug}
-										</span>
-										{relation.bidirectional && (
-											<span className="text-[10px] text-gray-400">↔ 双向</span>
-										)}
-										<button
-											type="button"
-											onClick={() => {
-												setFormData({
-													...formData,
-													relations: formData.relations.filter(
-														(_, i) => i !== index,
-													),
-												});
-											}}
-											className="text-gray-400 hover:text-red-500 transition-colors"
-										>
-											<X size={16} />
-										</button>
-									</div>
-								))}
-							</div>
-						)}
-
-						<div className="flex flex-col sm:flex-row gap-3 p-4 bg-brand-cream/30 rounded-2xl border border-brand-cream">
-							<select
-								value={newRelation.type}
-								onChange={(e) =>
-									setNewRelation({
-										...newRelation,
-										type: e.target.value as WikiRelationType,
-									})
-								}
-								className="px-4 py-2 bg-white rounded-xl border border-gray-200 text-sm"
-							>
-								<option value="related_person">相关人物</option>
-								<option value="work_relation">作品关联</option>
-								<option value="timeline_relation">时间线关联</option>
-								<option value="custom">自定义关系</option>
-							</select>
-							<div ref={relationSearchRef} className="relative flex-1">
-								<input
-									type="text"
-									value={newRelation.targetSlug}
-									onChange={(e) => {
-										handleRelationInputChange(e.target.value);
-									}}
-									placeholder="输入 [[页面标题]] 或直接输入 slug"
-									className="w-full px-4 py-2 bg-white rounded-xl border border-gray-200 text-sm"
-									onKeyDown={(e) => {
-										if (!showRelationDropdown) return;
-										if (e.key === "ArrowDown") {
-											e.preventDefault();
-											setRelationSelectedIndex((prev) =>
-												Math.min(prev + 1, relationSearchResults.length - 1),
-											);
-										} else if (e.key === "ArrowUp") {
-											e.preventDefault();
-											setRelationSelectedIndex((prev) =>
-												Math.max(prev - 1, -1),
-											);
-										} else if (
-											e.key === "Enter" &&
-											relationSelectedIndex >= 0
-										) {
-											e.preventDefault();
-											const selected =
-												relationSearchResults[relationSelectedIndex];
-											setNewRelation({
-												...newRelation,
-												targetSlug: selected.id || "",
-											});
-											setShowRelationDropdown(false);
-										} else if (e.key === "Escape") {
-											setShowRelationDropdown(false);
-										}
-									}}
-								/>
-								<AnimatePresence>
-									{showRelationDropdown && (
-										<motion.div
-											initial={{ opacity: 0, y: -4 }}
-											animate={{ opacity: 1, y: 0 }}
-											exit={{ opacity: 0, y: -4 }}
-											className="absolute z-50 mt-1 w-full bg-white rounded-xl border border-gray-200 shadow-lg max-h-60 overflow-auto"
-										>
-											<div className="px-3 py-2 text-xs text-gray-500 border-b border-gray-100 bg-gray-50">
-												直接输入 <code className="bg-gray-200 px-1 rounded">[[标题]]</code> 格式更快
-											</div>
-											{relationSearchLoading ? (
-												<div className="px-4 py-2 text-sm text-gray-500">
-													搜索中...
-												</div>
-											) : relationSearchResults.length === 0 ? (
-												<div className="px-4 py-2 text-sm text-gray-500">
-													未找到相关页面，请尝试输入 [[页面标题]]
-												</div>
-											) : (
-												relationSearchResults.map((result, idx) => (
-													<div
-														key={result.id}
-														className={`px-4 py-2 cursor-pointer ${idx === relationSelectedIndex ? "bg-brand-primary/10" : "hover:bg-gray-50"}`}
-														onClick={() => {
-															setNewRelation({
-																...newRelation,
-																targetSlug: result.id || "",
-															});
-															setShowRelationDropdown(false);
-														}}
-														onMouseEnter={() => setRelationSelectedIndex(idx)}
-													>
-														<div className="text-sm font-medium text-brand-olive">
-															{result.text}
-														</div>
-														<div className="text-xs text-gray-500 truncate flex items-center gap-2">
-															<span className="bg-gray-100 px-1.5 py-0.5 rounded text-[10px]">
-																{result.id}
-															</span>
-															<span>{result.subtext}</span>
-														</div>
-													</div>
-												))
-											)}
-										</motion.div>
-									)}
-								</AnimatePresence>
-							</div>
-							<input
-								type="text"
-								value={newRelation.label || ""}
-								onChange={(e) =>
-									setNewRelation({ ...newRelation, label: e.target.value })
-								}
-								placeholder="显示名称 (可选)"
-								className="flex-1 px-4 py-2 bg-white rounded-xl border border-gray-200 text-sm"
-							/>
-							<label className="flex items-center gap-2 text-xs text-gray-500 whitespace-nowrap">
-								<input
-									type="checkbox"
-									checked={newRelation.bidirectional}
-									onChange={(e) =>
-										setNewRelation({
-											...newRelation,
-											bidirectional: e.target.checked,
-										})
-									}
-									className="rounded"
-								/>
-								双向关联
-							</label>
-						</div>
-					</div>
-
-					<div className="pt-8 flex flex-wrap justify-end gap-3">
-						<button
-							type="button"
-							onClick={() => handleSubmit("draft")}
-							disabled={Boolean(savingMode)}
-							className="px-8 py-4 bg-gray-100 text-gray-700 rounded-full font-bold hover:bg-gray-200 transition-all flex items-center gap-2 disabled:opacity-50"
-						>
-							<Save size={18} />{" "}
-							{savingMode === "draft" ? "保存中..." : "保存草稿"}
-						</button>
-						<button
-							type="submit"
-							disabled={Boolean(savingMode)}
-							className="px-12 py-4 bg-brand-olive text-white rounded-full font-bold hover:bg-brand-olive/90 hover:scale-105 active:scale-95 transition-all shadow-lg flex items-center gap-2 disabled:opacity-50"
-						>
-							<Sparkles size={18} />{" "}
-							{savingMode === "pending" ? "提交中..." : "提交审核"}
-						</button>
-					</div>
-				</form>
-			</div>
-		</div>
-	);
-};
+const WikiEditor = WikiEditorComponent;
 
 // --- Wiki History Component ---
 const WikiHistory = () => {
