+++ src/types/common.ts
@@ -1,0 +1,13 @@
+export type ContentStatus = 'draft' | 'pending' | 'published' | 'rejected';
+
+export type FavoriteTargetType = 'wiki' | 'post' | 'music';
+
+export type AdminRole = 'user' | 'admin' | 'super_admin';
+
+export type UserStatus = 'active' | 'banned';
+
+export type Platform = 'netease' | 'tencent' | 'kugou' | 'baidu' | 'kuwo';
+
+export type ViewMode = 'large' | 'medium' | 'small' | 'list';
+
+export type ThemeName = 'default' | 'academy';
